docs: link element to partials (#519)
* docs: link element to partials

The link-element page is placed in the main table of contents within For Developers > Wallet linking and also within API docs > HTML DOM API. The reason is that this page is appropriate in both places.

However, since it's the same page, but with different table of contents titles, there's a navigation issue. If you click Next at the bottom of Wallet linking > Webpage (HTML), it takes you to the Feed page. If you click Next at the bottom of HTML DOM API > Monetization link element, it ALSO takes you to Feed instead of taking you to Monetization interfaces. If you use Next to navigate, you end up stuck in a loop.

I've made the content of the link-element page into a partials so that the partials can be embedded within two different pages, allowing the navigation to behave as expected.

* docs: fix broken links

Removed links to #web-monetization-agent from the -multilink and -iframes partials since they were breaking.