docs: nav updates (#521)
specified "HTML" page in link element overview partial, changed page title of monetization link element page version that is in the wallet linking section. changed the "next" link at the bottom of the publishers tools page from the spec to the compatible wallets page, because the spec opens in a new window. this change at least keeps the reader in the docs.